M2 cleanup: retire the orphaned ArgumentVoiceMap + dead voices-gate
The old per-section voice-dispute map is fully superseded by the statement spine.
Delete ArgumentVoiceMap.tsx (597 lines, app-unused since PR1); move its display
types (ArgumentVoice/ArgumentEdge/ArgumentVoicesData) to typing/voices.ts beside
deriveVoiceEdges. Trim useVoicesGate to just profile() (the showVoiceMap/
showFallback/mapEligible binders + their profile.ts imports were dead); the pure
gate fns stay in profile.ts with their tests. Prune the now-obsolete ArgumentVoiceMap
RTL describe block (kept the live ArgumentFlowGraph one).
